Make interpreter work on newer versions of Java
In previous versions of Java (or perhaps 32-bit, rather than 64-bit
versions) hashCode() always appeared to return positive numbers. This is
not guaranteed, however, and this modification makes the code run
properly on implementations that return negative numbers.

The Object.hashCode() is used to populate a method cache, so this
modification does not affect the image, just the virtual machine.