Add custom image serialization
SmallWorld originally depended on Java's built-in ObjectInputStream and
ObjectOutputStream. This output format is fragile, because it is very
sensitive to object layout. To make the image format less brittle to
changes in the java code, this uses a custom and extremely basic
implementation of an object input stream, custom to the interpreter.

This has the potential advantage that the image can now be used outside
of Java, because the image doesn't depend on any java depdencies. The
pool of objects is represented as big-endian integers and references to
other objects.

The current implementation is a bit fragile - changing the order of the
reading or writing the roots will break the format.  This will be fixed
in the future, but it paves the way to something more important -
changing the java code without fear of breaking the image.