Fix thread preemption issues
Thread.stop() is deprecated because it is not safe. It was being used
when waiting on a semaphore (for example, Window>>question: spawns a
thread, and the ActionListener wakes up the thread, but the cancel
button simply kills the thread).

Instead, the threads now call Thread.interrupt(), which causes the
semaphore to (correctly) throw an InterruptedException. When the
interpreter encounters this, it returns nil and the execution stops.

One other tweak in this CL was to avoid the static debug flag. Now, the
entire codebase compiles without warnings!