First steps toward a SmallWorld text mode REPL.
This is a clone of the GUI, but using the dummy UI Factory. It
successfully runs the doIt handler, which menas a reasonably large
amount of the interpreter can work properly both without a GUI (and
also, it seems, without threads).

This is an incremental step twoard an actual interactive REPL (the
strings being run are hard coded in the main method), but should also
allow for a potential port of this to something like TeaVM or another
transpiler for the Web.