Add experimental conversion to JavaScript.
This leverages TeaVM to trasnspile the Java code into JavaScript.
Surprisingly, simple tests of this seem to work.

This code is very hacky, but it is possible to run a super simple
version of SmallWorld in a Web browser, using only JavaScript. There is
no GUI, but the doIts successfully print to the JavaScript console.

See more discussion in experimental/README.md