Make experimental js code depend on SmallWorld.jar
Previously, all the code in the main repository (smallworld.*) was
cloned into experimental, so that TeaVM could recompile it. This was
unnecessary, and now the experimental TeaVM version just has the few
files it needs to make the javascript based REPL.

This depends on using gradle to build the jar, and then using maven
in the experimental directory to build the js bundle.

However, experimental now has much less code.